work on migrating on microk8s

+++ scripts/cluster.sh
@@ -37,6 +37,7 @@ function cluster_init() {
 
   apply_nginx_ingress
   apply_cert_manager
+  apply_storage
 
   sleep 2
 
@@ -57,8 +58,17 @@ function apply_nginx_ingress() {
   # k3s : 'cloud'
   # kind : 'kind'
   # kubectl apply -f https://raw.githubusercontent.com/kubernetes/ingress-nginx/controller-v1.1.2/deploy/static/provider/cloud/deploy.yaml
+  if [ "${CLUSTER_RUNTIME}" == "kind" ] || [ "${CLUSTER_RUNTIME}" == "k3s" ]; then
+    kubectl apply -f kube/ingress-nginx-"${CLUSTER_RUNTIME}".yaml
+  elif [ "${CLUSTER_RUNTIME}" == "microk8s" ]; then
+    microk8s enable ingress
 
-  kubectl apply -f kube/ingress-nginx-"${CLUSTER_RUNTIME}".yaml
+    kubectl -n ingress patch daemonset.apps/nginx-ingress-microk8s-controller --patch\
+  "{\"spec\":{\"template\":{\"spec\":{\"containers\":[{\"name\":\"nginx-ingress-microk8s\",\"ports\":[{\"name\":\"http\",\"hostPort\":${NGINX_HTTP_PORT},\"containerPort\":80},{\"name\":\"https\",\"hostPort\":${NGINX_HTTPS_PORT},\"containerPort\":443}]}]}}}}"
+  else
+    echo "Unknown CLUSTER_RUNTIME ${CLUSTER_RUNTIME}"
+    exit 1
+  fi
 
   pop_fn
 }
@@ -66,18 +76,31 @@ function apply_nginx_ingress() {
 function delete_nginx_ingress() {
   push_fn "Deleting ${CLUSTER_RUNTIME} ingress controller"
 
-  kubectl delete -f kube/ingress-nginx-"${CLUSTER_RUNTIME}".yaml
-
+  if [ "${CLUSTER_RUNTIME}" == "kind" ] || [ "${CLUSTER_RUNTIME}" == "k3s" ]; then
+    kubectl delete -f kube/ingress-nginx-"${CLUSTER_RUNTIME}".yaml
+  elif [ "${CLUSTER_RUNTIME}" == "microk8s" ]; then
+    microk8s disable ingress
+  else
+    echo "Unknown CLUSTER_RUNTIME ${CLUSTER_RUNTIME}"
+    exit 1
+  fi
   pop_fn
 }
 
 function wait_for_nginx_ingress() {
   push_fn "Waiting for ingress controller"
 
-  kubectl wait --namespace ingress-nginx \
-    --for=condition=ready pod \
-    --selector=app.kubernetes.io/component=controller \
-    --timeout=2m
+  if [ "${CLUSTER_RUNTIME}" == "kind" ] || [ "${CLUSTER_RUNTIME}" == "k3s" ]; then
+    kubectl wait --namespace ingress-nginx \
+      --for=condition=ready pod \
+      --selector=app.kubernetes.io/component=controller \
+      --timeout=2m
+  elif [ "${CLUSTER_RUNTIME}" == "microk8s" ]; then
+    :
+  else
+    echo "Unknown CLUSTER_RUNTIME ${CLUSTER_RUNTIME}"
+    exit 1
+  fi
 
   pop_fn
 }
@@ -86,8 +109,14 @@ function apply_cert_manager() {
   push_fn "Launching cert-manager"
 
   # Install cert-manager to manage TLS certificates
-  kubectl apply -f https://github.com/jetstack/cert-manager/releases/download/v1.6.1/cert-manager.yaml
-
+  if [ "${CLUSTER_RUNTIME}" == "kind" ] || [ "${CLUSTER_RUNTIME}" == "k3s" ]; then
+    kubectl apply -f https://github.com/jetstack/cert-manager/releases/download/v1.6.1/cert-manager.yaml
+  elif [ "${CLUSTER_RUNTIME}" == "microk8s" ]; then
+    microk8s enable cert-manager
+  else
+    echo "Unknown CLUSTER_RUNTIME ${CLUSTER_RUNTIME}"
+    exit 1
+  fi
   pop_fn
 }
 
@@ -95,7 +124,14 @@ function delete_cert_manager() {
   push_fn "Deleting cert-manager"
 
   # Remove cert-manager
-  kubectl delete -f https://github.com/jetstack/cert-manager/releases/download/v1.6.1/cert-manager.yaml
+  if [ "${CLUSTER_RUNTIME}" == "kind" ] || [ "${CLUSTER_RUNTIME}" == "k3s" ]; then
+    kubectl delete -f https://github.com/jetstack/cert-manager/releases/download/v1.6.1/cert-manager.yaml
+  elif [ "${CLUSTER_RUNTIME}" == "microk8s" ]; then
+    microk8s disable cert-manager
+  else
+    echo "Unknown CLUSTER_RUNTIME ${CLUSTER_RUNTIME}"
+    exit 1
+  fi
 
   pop_fn
 }
@@ -103,9 +139,59 @@ function delete_cert_manager() {
 function wait_for_cert_manager() {
   push_fn "Waiting for cert-manager"
 
-  kubectl -n cert-manager rollout status deploy/cert-manager
-  kubectl -n cert-manager rollout status deploy/cert-manager-cainjector
-  kubectl -n cert-manager rollout status deploy/cert-manager-webhook
+  if [ "${CLUSTER_RUNTIME}" == "kind" ] || [ "${CLUSTER_RUNTIME}" == "k3s" ]; then
+    kubectl -n cert-manager rollout status deploy/cert-manager
+    kubectl -n cert-manager rollout status deploy/cert-manager-cainjector
+    kubectl -n cert-manager rollout status deploy/cert-manager-webhook
+  elif [ "${CLUSTER_RUNTIME}" == "microk8s" ]; then
+    :
+  else
+    echo "Unknown CLUSTER_RUNTIME ${CLUSTER_RUNTIME}"
+    exit 1
+  fi
+  pop_fn
+}
+
+function apply_storage() {
+  push_fn "Launching storage"
+
+  if [ "${CLUSTER_RUNTIME}" == "kind" ] || [ "${CLUSTER_RUNTIME}" == "k3s" ]; then
+    :
+  elif [ "${CLUSTER_RUNTIME}" == "microk8s" ]; then
+    cat <<EOF | kubectl apply -f -
+apiVersion: storage.k8s.io/v1
+kind: StorageClass
+metadata:
+  name: local-storage
+provisioner: kubernetes.io/no-provisioner
+volumeBindingMode: WaitForFirstConsumer
+EOF
+  else
+    echo "Unknown CLUSTER_RUNTIME ${CLUSTER_RUNTIME}"
+    exit 1
+  fi
+
+  pop_fn
+}
+
+function delete_storage() {
+  push_fn "Launching storage"
+
+  if [ "${CLUSTER_RUNTIME}" == "kind" ] || [ "${CLUSTER_RUNTIME}" == "k3s" ]; then
+    :
+  elif [ "${CLUSTER_RUNTIME}" == "microk8s" ]; then
+    cat <<EOF | kubectl delete -f -
+apiVersion: storage.k8s.io/v1
+kind: StorageClass
+metadata:
+  name: local-storage
+provisioner: kubernetes.io/no-provisioner
+volumeBindingMode: WaitForFirstConsumer
+EOF
+  else
+    echo "Unknown CLUSTER_RUNTIME ${CLUSTER_RUNTIME}"
+    exit 1
+  fi
 
   pop_fn
 }
@@ -113,4 +199,5 @@ function wait_for_cert_manager() {
 function cluster_clean() {
   delete_nginx_ingress
   delete_cert_manager
+  delete_storage
 }
